Improve setting and getting volatile attributes.  #3353.
This update filters setting of attributes based on some prefix, defaulting
to "Chirp".  Any attribute set that does not start with the prefix will be
ignored.

This also adds the get_job_attr_volatile command which retrieves an attribute
from the starter's copy of the ad.

Security settings are firmed up to make sure they reflect what is
written in the corresponding ticket.