Gridmanager can crash on certain replies from the EC2 GAHP #3701
When a command fails, the EC2 GAHP usually returns two error strings
detailing the failure. In some cases, it doesn't return these strings.
Many places in the EC2Job object of the gridmanager assume these strings
are always present on a failure, and will dereference a NULL pointer if
they are missing.

Add checks to avoid the dereferences when the string pointer is NULL.