Try and fail to get condor_dagman_metrics_reporter to build on windows #3532
isspace redefined to is_space in b963f58. This seems like a bad idea.

Check for curl to build metrics reporter. This failed on Solaris in
batlab, and seems to cause linking problems on Windows.

Do not build dagman_metrics_reporter on windows (see above).  Windows
linker is not the most helpful tool in the world, and debugging is
difficult since NMI deletes everything on error.