Fix printing of nested ads in ClassAdUnParser for old ad syntax. #3772
When unparsing an ad in old ClassAd syntax, any nested ads should be
delimited in new syntax (enclose in square brackets, seperate attributes
with semicolon), though other aspects of the old syntax should be
retained (string escaping, meta-equality operators).

We add a new SetOldClassAd(bool, bool) method for use by Condor, which
usually uses the unparser to print individual attribute values. In that
case, classads should always be delimited in the new syntax.

Due to object packing and alignment, the added boolean datamember
oldClassAdValue shouldn't change the size or layout of ClassAdUnParser,
so we shouldn't need to bump the SO version for this change.