Make a CondorVersionInfo from a set of version integers. #3856
Add a new constructor for CondorVersionInfo that takes the broken-down
components of a Condor version: integers for the components of the
version and an optional char* for the build date and other trailing
info.
This is useful for the gridmanager and ft-gahp to feed default peer
versions into the FileTransfer object for backwards-compatibility.