Add CONDOR_VERSION command to gridmanager. #3856
The file transfer protocol changed in 8.1.0. Newer daemons/tools need to
know when they're talking to an older peer or else transfers will fail.
Since the gridmanager and ft-gahp reside in different machines, it's
very possible that they will be different versions.
Thus, the ft-gahp and gridmanager must exchange version strings and feed
those strings into the FileTransfer object.
The gridmanager will always issue the CONDOR_VERSION command to a GAHP
that supports it.