Create a CEDAR session between the gridmanager and the ft-gahp. #3536
Create new command CREATE_CONDOR_SECURITY_SESSION for the ft-gahp.
The gridmanager creates a security session and uses this gahp command to
have the ft-gahp create the same session. Then, all file transfers use
this session when setting up the CEDAR connection.
This allows the ft-gahp to work without prior manual configuration of
the ft-gahp and gridmanager's security policy in ways that could
conflict the existing policy.

The security session is formatted the same way as when a daemon passes a
security session to its daemon-core child process.
When the GAHP exits, the gridmanager removes the security session.

Since the new command contains sensitive security information, we
obscure the information in daemon logs by default and add the new config
param GAHP_DEBUG_HIDE_SENSITIVE_DATA to allow the data to be unobscured
when debugging.