Print reals in ClassAds in a more human-readable format. #3928
ClassAds' standard way of printing floating point values, always using
scientific notation, makes it harder for humans to read.
We can make values easier to read with maintaining full precision by
using the %G printf format instead of %E. It uses scientific notation
only for very large or very small values. It also trims unnecessary
zeros after the decimal point.

We use this different format only when SetOldClassAd() is enabled in the
unparser, as it breaks the canonical representation of ClassAds in their
native form, as described in the specification.