ClassAd quantize() can fail when in a cached expression. #3967
When the second argument to quantize() is a list, we call Evaluate() on
each element in the list. The variant of Evaluate() we use looks up the
parent scope of the ExprTree for resolving any attribute references. If
the ExprTree is part of a cached expression, that scope will be NULL,
leading evaluation to fail. Instead, we should use the Evaluate()
variant that takes an EvalState, which we have handy.