Fixes needed to allow build with HAVE_EXT_POSTGRESQL
This commit implements changes needed to allow the software to build
with HAVE_EXT_POSTGRESQL defined. The DB components use sprintf in
MyString rather than formatstr, and to limit the scope of the changes
I added an "sprintf" function as a duplicate of the formatstr
function, rather than rewriting dozens of sprintf references in
multiple files.