Allow explicit use of TCP updates for CONDOR_VIEW_HOST. #3986
Updates to the CONDOR_VIEW_HOST will use TCP if we need to use CCB or
shared point to reach it. But there was no way to explicitly request use
of TCP otherwise.
Now, UPDATE_COLLECTOR_WITH_TCP and TCP_UPDATE_COLLECTORS are respected
when sending updates to the CONDOR_VIEW_HOST.