New parameter for setting user execdir permissions #4208
User can set JOB_EXECDIR_PERMISSIONS variable to one of:
    * user (default)
    * group
    * world
that will make the job's scratch directory have permissions 0700, 0750,
or 0755 respectively.

At the RACF the PHENIX experiment relied on the the directory being
at least group-readable, since the setuid copy-as-user program for
transferring output on behalf of another user needs to read files in
this directory.

Conflicts:
	src/condor_utils/param_info.in