Use OAuth 2.0 service to get necessary credentials. #3952
Instead of assuming the auth file has a bare access token, we now expect
to have a refresh token and related data, formatted in JSON as generated
by the gcutil command-line tool. We use that information to obtain an
access token from Google's OAuth 2.0 service. When that token is close
to expiring, we obtain another one.