Wait for asynchronous requests to finish in gce_gahp. #3952
Some requests in GCE return immediately, before the operation has taken
place. The response includes a handle that can be used to check the
status of the operation later. Instance creation and deletion work this
way. The gce_gahp now waits for these operations to complete, polling
their status periodically.
One side effect of waiting is that we can now return the instance id in
the results of GCE_INSTANCE_INSERT commands.