Add python bindings for the event log reader API. #4071
This exposes a new function, read_events, in the htcondor python module.
Given a file object, it returns an iterator of ClassAds corresponding
to the HTCondor events in that log file.

The functionality is purposely kept simple - locking logs, writing to
logs, and managing missed events / log rotations are left as future work.

Note the implementation sticks to the public API / ABI of the event log reader
even though this API has usability issues due to its forward dec'l of
compat_classad.  Theoretically, this could be compiled separately from HTCondor.

This commit covers test cases and documentation.