Fix indenting, no other changes
Normally fixing indenting is a bad idea.  But in this case the file was
a mess.  Different parts of the file expected a 4-space tab and an
8-space tab, sometimes within a single function.  Lines mixed tabs and
spaces, sometimes with the spaces between tabs.  In a few places, a
3-space indent was being used.  For my own sanity, everything has mashed
flat to use strictly tabs for indenting; it should now work fine with
any width of tab.