New DEPRECATED macro generates compiler warnings
When a function declaration is wrapped in DEPRECATED() (and you #include
"deprecated.h"), any uses of that function will generate compiler
warnings in Visual Studio and GCC.  Currently just a single function
(RemoteResource::setStartdInfo) is flagged as such; its comment claimed
it was.  Added with the goal of using in the IPv4/IPv6 mixed-mode work.