Fix filename collision when NorduGrid ARC submits into HTCondor. #4017
When the gridmanager submits a job to nordugrid, it names the stdout/err
files _condor_stdout/err. When a Condor vanilla universe job uses file
transfer, it also names the stdout/err files _condor_stdout/err.
NorduGrid ARC wraps the job with a script that explicitly captures the
job's stdout/err using the filenames given by the submitter.
Put all of these together and the job's stdout/err gets mingled with
that of the wrapper script and not returned to the submitter.

Change the names used by the gridmanager to avoid this. We include a
sanitized version of the GlobalJobId to avoid possible collisions in the
future.