Do not write unparseable values into classad collection logs. #3616
Before writing an attribute value to the log (aka job_queue.log), make
certain the value can be parsed. This is done efficiently because the
parsing is only done once when the log event is created. The parse
tree is then stored in ram and reused when the log is played. If the
value fails to parse, the value is changed to UNDEFINED.