Support setting of metadata in the gce_gahp. #3952
Metadata can be set directly in the submit file or in a seperate file,
similar to EC2 user data. The formatting is different, as GCE metadata
is ascii key/value pairs.
Metadata given in the submit file via gce_metadata is formatted
identically to the job environment attribute.
If a seperate file is used, it must contain one key/value pair per line,
with '=' or ':' as the seperator, and all whitespace is not ignored.