Fix crashes on F19 with optimization an USE_CLONE...=true #4009
Compiler optimization was defeating our detection of
the direction of stack growth.  Given that there's no truely
portable way to do this, and the clone man page tells us
the only architecture where stack grows up is HP-PA, and we are
unlikely to ever run on linux on hppa, just hardcode this as
stack_grows_down.