Have CEDAR manage the TCP_KEEPALIVE settings for ReliSocks. #4122
Keepalive is set to 6 minutes.  360 seconds was picked because it's
slightly greater than the default UPDATE_INTERVALs for the starter
and updating the collector via TCP, but still small enough that
a dead TCP socket should be detected before the default job lease
duration (currently 20 minutes) expires.

Because it is greater than the default UPDATE_INTERVALs for our
long-lived TCP connections, it is expected this change will only
cause additional TCP traffic when there are real TCP issues.

TCP keepalive has been enabled for about 8 years, but likely it was
never used because most sites have a timeout of 2 hours -- meaning
that the other side probably had long given up on the connection.