Disable almost all tests on Darwin temporarily. #4157
There's a new bug on master that causes daemons to constantly try to
accept bogus connections, chewing up as much cpu as possible.
When enough daemons are hit by this in a test suite run, it can bring
down the machine. Disable all but one test until we fix the problem.