Change daemoncore handler tables into dynamically-resizeable arrays. #4206
The arrays comTable, sigTable, and reapTable were statically allocated
with many more entries than are typically used. Now they dynamically
resizeable with a small initial size.
One potential drawback is that lookup time is now linear. If that is an
issue, we could keep the entries sorted and do a binary search for
lookup.