Don't call reapers immediately after select() in the main event loop. #4210
It's a bad idea to call reaper handlers after select() but before
looking up handlers for any FDs that are ready.
Reapers can close FDs and open new FDs that get the same numbers. We can
then end up calling a handler for an FD that's not actually ready.
This is happening with the fetch work hook in the startd, causing the
startd to do a blocking read on an FD that has no data.