Use updated gridftp client library in nordugrid_gahp. #4243
Current versions of the Globus gridftp client library include the
explicit CWD command we had to add to support job submission to
NorduGrid ARC. Start using the current official library instead of our
hacked copy of an older version.
One benefit is that this allows us to support IPv6 in the
nordugrid_gahp.