Eviction while ssh_to_job active crashes the starter. #4251
Allow the SSHDProc to be responsible for deleting the job ad it's given.
Before, it was given a pointer to an ad on the stack. Any subsequent
access to the ad would cause a crach.
One instance where this happens is during shutdown, when the SSHDProc
looks up which signal to send to the sshd process.