Update BOINC_QUERY_BATCHES gahp command syntax. #3540
The updated command takes a time of last query and returns the time of
the current query. This allows us to ask for only jobs that have changed
status since the last time we asked.
For now, we just ask for all changes since the beginning of time (the
old behavior).