Grow buffer to be one byte larger than what we need. #4273
It's apparently impossible to seek() to the end of a buffer due to
an off-by-one error in Buffer::seek(); this was never trigger-able before
because of the fixed size of buffers.

Due to this off-by-one, responses filling the entire buffer are rejected.

I hesitate to fix Buffer::seek because it's not clear how to check that
nothing else in CEDAR depends on this precise behavior.