Allow the user to disable spooling of output files on evictions. #4292
This allows the user to recover output files if the job is removed
from the queue.  The original use case is for CMS, where jobs that go
over their memory limits are removed via PeriodicRemove and users would
like to be able to read the job's log files to figure out why so much
memory was used.