Undo change in exit code checks for file transfer child processes. #4328
When a process created with Create_Thread() exits, we terminate the
process by exec'ing /bin/true or /bin/false. This means any non-zero
return value of the child is translated into the exit code of
/bin/false. On Solaris, this is 255, not 1.

For 8.0, we'll rely on the final status update to catch failures that
result in exiting with a non-zero status.