Fix some build problems with the CREAM external 1.14.0. #4392
These issues caused communication with CREAM servers to fail.

First, re-apply a patch for glite-lbjp-common-gss, which gives a bogus
buffer length to gss_import_cred().

Second, glite-lbjp-common-gsoap-plugin needs to be built as C++ code and
we need to link with libgsoapssl++.a

Third, reorganize the CREAM-related externals that that they don't all
install into the installation directory for glite-ce-cream-client-api-c.