Improve directory execute permission check in access_euid(). #4402
When checking whether the current euid has execute permission on a
directory, attempt the operation instead of manually examining bits in
the inode. We do this for read and write permissions, and it takes into
account all ways in which the operation may be permitted, including
supplemental groups and POSIX ACLs.