Lean on OS for access() with euid when possible. #4402
access() does file permission checks using the process's real uid. We
want to do checks with the effective uid, so we wrote our own
access_euid(). Recent versions of linux now provide euidaccess(), which
does what we want and should do it better than us. Use it when it's
available.