Ensure thread safety for Globus and OpenSSL in cream_gahp. #4406
Since we're creating multiple worker threads in the cream_gahp, we need
to call certain initialization functions for Globus and gsoap in the
main thread before any new threads are created. Otherwise, OpenSSL and
Globus functions/data are accessed in a non-thread-safe way.
When under load, this results in frequent crashes.