Relative proxy path doesn't work in submit file. #4415
When checking for the validity of the job's proxy, a relative path was
made relative to the CWD of condor_submit, but the path name place in
the job ad was relative to the job's initialdir.
If the CWD and initialdir are different, then either job submission would
fail or execution would fail.
Now, the proxy path is always relative to initialdir.