Make parent-child security session in CEDAR usable. #4405
In order for the pre-created security session between parent and child
daemons to be usable, the child needs to know the list of command
integers that the session can be used with. We now pass that in the
serialized session string passed from parent to child.
The child also needs to include the parent's sinful string in the
session structure it creates from the inherited serialized string.