cream_gahp fails at startup due to lt_dlopen() search path. #4440
Globus uses lt_dlopen() to load its thread library,
globus_thread_pthread. This function appears to ignore the program's
RPATH, so it doesn't find the copy of the library that we include in UW
builds of Condor.
We can work around this problem by loading the library with dlopen()
before Globus calls lt_dlopen().