Add support for PR_SET_NO_NEW_PRIVS (#4443).
This adds a new parameter, JOB_NO_NEW_PRIVS (defaults to False)
which causes prctl(PR_SET_NO_NEW_PRIVS, 1) to be invoked prior
to exec'ing the job.  When this prctl flag is set, the job and
its dependent processes will be unable to acquire new privileges
(for example, through the use of setuid binaries, a common attack
vector).

This defaults to off because there are some use cases (specifically,
glexec) that require setuid binaries to function.