When available, use epoll for idle CCB sockets.
CCB may manage a huge number of sockets - many more than
what is considered possible to maintain inside DaemonCore (as
DC is currently based on select() and may have other overhead).

The current approach is to periodically poll each socket, one
at a time.  This approach remains on platforms without epoll,
but may occur up to a 20 second wait for a response on busy
CCB instances.

Now, epoll is used on supported platforms; this allows CCB to
be more responsive when targets send responses.  The epoll FD
is managed by DC; when it fires, we call epoll directly to drain
as many sockets as possible.