Look in headers, not libc, for fdatasync.  #4283
Mac OS X is a curious platform.  Its libc provides fdatasync as
a symbol we can link against, but the headers we use don't actually
define fdatasync.  Hence, the build is failing because the CMake
test succeeds though condor_fsync.cpp doesn't compile.

So, we switch the cmake test from looking at libc to the unistd.h
header for fdatasync.  This will disable use of fdatasync on Mac
OS X - but fix the build.