Fix file descriptor leak in condor_shared_port. #4456
SharedPortState assumes DaemonCore will clean up the ReliSock and FD
used to pass a socket FD to a daemon. This only happens when the
ReliSock was registered with DaemonCore (when SharedPortState is in
non-blocking mode). If SharedPortState hasn't registered the ReliSock
(mainly in blocking mode), it needs to do the cleanup.

The main place this cleanup wasn't happening was when the shared-port
daemon sent a DC_CHILDALIVE message to the master.