Enforce Globus thread model "none". #4464
If the Globus thread model is not "none", then Globus will use
lt_dlopen() to load a Globus library with routines for the given thread
library. This is likely to fail for UW builds of Condor, as lt_dlopen()
ignores the RPATH, which is how we locate the Globus libraries we ship
with Condor.
The thread model can be set via GLOBUS_THREAD_MODEL in the environment.
Override that with an explicit call to globus_thread_set_model().