Fix how arguments are passed to remote_gahp by the gridmanager. #4434
Now, for grid batch jobs that want to use remote_gahp, the options
included in the GridResource attribute are placed before the name of the
remote gahp binary in the argument list for remote_gahp.
These options are meant for remote_gahp, so this makes more sense than
the previous ordering.