Fixups based on batlab tests. #3813
For the most part, disable tests which require shared port to
be turned off to function (for example, if they want to detect
the TCP port used by a daemon and cannot parse a sinful string).

Notably, this also defaults shared port to FALSE for Win32.  This
is because there are some suspicious threading issues in batlab
tests which I don't have the expertise to solve.