Fix bug setting socket buffer sizes on newer linux kernels. #4489
The problem was on the 3.14 kernel, the kernel doesn't support
setting the send buffer less than about 4k bytes, so our initial
attempts to set it to 1k and 2k both result in setting it to 4k.
Condor notices that the size isn't going up, and stops trying
well short of the requested size.