Sock::bind can now often correctly identify protocol
The non-condor_protocol version of Sock::bind assumed IPv4, unless IPv6
was enabled in which case it assumed IPv6.  Now if _who has been set, it
identifies the correct protocol from that.  This function remains a
temporary one, ideally everyone should move to the Sock::bind that takes
a condor_protocol.

Also eliminated a useless test and flagged a few code paths that should
be removed eventually.