Do job proxy operations in a child process of the shadow on linux. #4483
We want to avoid loading the GSI libraries in the shadow process. Thus,
when we need to read the job proxy, fork a quick child process to do the
actual operations. The results of the operation are sent to the main
shadow process via a pipe.
File transfer is already done in a child process, so no change required
there.