Add very basic debugging tool for JOB_ROUTER. #4569 condor_job_router_tool
This new tool can read the job router configuration and the dump the table of routes.
It can also read job ad's from a file and match them to routes, printing
the first matching route for each job that has a match.

This tool works by invoking the actual JobRouter class, but hijacking it's
inputs and outputs - consequently, the formatting of the output is currently
the same as the dprintf output of the condor_job_router. This produces a tool with
minimal code change to condor_job_router proper so that the patch is suitable for the stable branch.