Addition of the log reader API for python. #4588
This is an iterator-based API for reading through a ClassAd log
(such as written by the schedd).  The parser and probe is reused
from the existing ClassAdLog* API, but the control flow was rewritten
to match the inverted, iterator-based style preferred in python.

The API provides inotify support for clean integration within
event loops and several options to decide whether to block (and
for how long).

This commit includes unit tests but not documentation.