Rotation of shared daemon log can result in writing to old file #3106
When a daemon forks off a helper process on unix, the child process now
reopens the log file on every write and doesn't perform log rotation.
This fixes a bug where both parent and child will rotate the log, and
one ends up continuing to write to the rotated file.