Allow sub-collector queries.  #4674
This functionality, inspired by "condor_q -g", allows the condor_status
(and anything else used by CollectorList) to first query the collector
for a list of child collectors, then performs the query sequentially
against the child collectors.

This functionality is controlled by COLLECTOR_QUERY_EXPR; the default
(false) keeps the previous behavior of querying the parent directly.
If, for example, this is set to "true", all child collectors and
the parent are queried; the parent is queried because it advertises
itself.

Similarly to how "condor_q -g" works, this has the following error
semantics: if there's at least one child collector that returns
successfully, the whole operation is marked as a success. Otherwise,
the operation fails.