Only send slot keepalives when there is no starter. #4491
Previously, the startd would constantly connect to the schedd
to send alive messages on the claim. Now if knob STARTER_HANDLES_ALIVES
is True in the schedd (the default), the slot lease is automatically
renewed while a starter/shadow exists, thus alive messages are only
sent when a slot is claimed/idle. This required changes to the starter
to notice when the syscall_sock to the shadow disappeared, and exit
if the syscall_sock is not reconnected within the slot lease period.