Allow fine-grained authorization for match-based auth. #4688
This allows the schedd to setup authorization (ALLOW_MATCH_AUTH/
DENY_MATCH_AUTH) for match authentication.  Instead of applying
logic to a CEDAR socket, it examines information from the slot ad.

In particular, it uses the slot ad's MyAddress and AuthenticatedIdentity
attribute; the latter is verified by the collector itself.

We have one piece of special logic - the token "self".  This allows
schedd to use match authentication if the slot ad's AuthenticatedIdentity
and the job ad's User are the same.