For versioned spooling protocol, use version from CEDAR socket. #4720
Previously, if DCSchedd was created from a sinful string, it would
assume the remote side was the same version as the local one (the
version was taken when it contacted the collector).  This broke any
protocol versioning.

Now, we query the CEDAR socket and assume the version we authenticated
with is the best source of this information.