File transfer failure with python bindings to 8.0 schedd. #4720
When the DCSchedd object is used to perform file transfers, if it doesn't
know the version of the schedd, it assumes the schedd is the same version
as itself. This can lead to failures if the schedd is old enough to not
know about file transfer protocol changes, such as the the one introduced
in 8.1.0.
Two cases where this can arise are with the Python bindings and when the
command lines tools are invoked with the -address option. In these cases,
the DCSchedd object is only given the sinful string of the schedd.

Once the DCSchedd object has connected to the schedd and performed
authentication, we can query the ReliSock for the schedd's version, which
was obtained during the security negotiation.