Nordugrid jobs held when exit info missing #4736
When the gridmanager queries the runtime stats of a completed nordugrid
job, it expects all the stats to be present, and will put the job on
hold if any are missing. It turns out most of these stats are optional.
Only throw an error if all stats are missing (it appears that
nordugrid-job-usedcputime will already be present, possibly with a bogus
value of 0).