This commit introduces the concept of a "Socket Manager", which manages the socket table and a corresponding selector. #189
Importantly, the SocketManager starts to provide more epoll-friendly
interfaces (iterating through ready sockets, not querying the status
of each).  The abstraction isn't complete - some internal details
(such as the position of the socket in the socket table) still leak
into DaemonCore.

Further, the pipes and async-signal handler are not incorporated into
the SocketManager design.  For now, we add/remove all other FDs in
each DC driver loop.  In the future, we'll need to upgrade it to a
"FD Manager" prior to switching to epoll.

The commit passes the basic smoke tests:
- Collector works
- Jobs run.
- Works with shared port / CCB

As a part of this squashed commit, generic_stats.h is slightly
refactored so it isn't dependent on other headers - it can be
included by itself.