Fix FileTransfer protocol mixup with output file plugins. #4739
When file transfer plugins are used for output files, transfer command
999 is used. This sends a classad with details about how the transfer
will occur, including a SubCommand attribute. The receiver
(DoDownload()) looks up the sub-command using the wrong attribute name
(Result).
Fortunately, this appears to be a mostly harmless bug currently. The
transfer will still proceed or fail as it should, but with worrying
error messages in the shadow log.