When building Globus ourselves, hard-core SO version of libraries. #4585
When we build Globus and VOMS, there's probably no library to check the
SO version of when cmake runs. The library will be created afterwards,
once we start compiling. Since we know which version of Globus/VOMS
we're using, we can hard-code the values. We just have to remember to
update them when necessary.