Delay Condor-C output file transfer until JobFinishedHookDone is set. #3379
When the schedd does file transfers, it accesses the job sandbox as the
condor user. Thus, the transfer of output files can fail if the sandbox
is still owned by the user. To guarantee that the job files are owned by
the condor user, the gridmanager must wait until JobFinishedHookDone is
set in the job ad.