Daemon log rotation disabled in schedd when using clone(). #4753
In the dprintf() code, the flag DebugRotateLog controls whether the
process should worry about rotating the log. We set this to false in
children fork()d from a daemon to avoid a race between parent and child
to rotate the log. Unfortunately, when clone() is used in
Create_Process(), changing DebugRotateLog in the child also changes it
in the parent. Consequently, the schedd stops rotating its log.
To fix, we set DebugRotateLog back to true when the child is done using
dprintf(), indicated via an existing function call.