Switch DC to epoll. #189
Using the central SockManager class, we now use epoll for all
registered sockets.  Note that select is still used for pipes - I
consider this OK because the daemon we are most concerned about
(the collector) keeps no pipes.

Note that, in this implementation, I disabled the special checks
of the superuser socket.  If we go this route, I plan on replacing
the superuser socket with a priority system for sockets.