Remove NotifyJobScheduler option from schedd's ACT_ON_JOBS command.
The schedd ignores this option for internal decisions to hold or remove
jobs (it always notifies), with a log message saying "this should not be
optional". It only respects it for requests coming from a client. None
of our current client code disables notification and doing so doesn't
make any sense.