API for querying remote daemon config. #4778
This adds a new class to the python bindings, RemoteParam.
A RemoteParam object (initialized with a ClassAd containing
at least the MyAddress attribute) behaves similarly to the
Param object and a python dictionary.

One difference is that, on access, the remote daemon is
queried for configuration instead of doing this at objection creation.