Record job ad at abort or terminate time into a file. #4784
This commit records the job ad at abort / terminate time before the
corresponding user log event is logged.  This guarantees that the
job's final(ish) ClassAd is available to a DAG's postjob script.

Note this is slightly different than what is found in the history
file - the job may not exit the queue after any given abort/terminate.

If nothing else, the JobStatus is still running when this code path
is reached.