Add special attribute My in new ClassAds, similar to self. #4782
In the ClassAds compatibility layer, we add and remove a My attribute
every time we do an evaluation or track references. The ClassAds library
already has a special 'self' attribute that refers to the current ad
(which we use to implement My). By having the library recognize My the
same as self, we eliminate these frequent ad moditifcations and many
compat layer methods can become const.

The recognition of My is contingent on the old ClassAds semantics flag
being set to true. This is now done via the new function
SetOldClassAdSemantics().
FlattenAndInline() no longer inlines entire classads.