Turn the TimerManager into a singleton. #4773
Previously, a TimerManager instance was a static member of the DaemonCore
class initialized in the static initializers.  The TimerManager was
not actually a singleton, but EXCEPT'd if the constructor was called
twice.

In #3137, condor_utils was switched to be statically compiled into
the condor_shadow.  Thus, the static initializers are run once when
the executable starts; if a library is dlopen'd (such as when
CLASSAD_USER_PYTHON_MODULES is set) that is linked against
libcondor_utils, the static initializer for daemon_core.cpp is run
again - and boom! - the TimerManager constructor EXCEPT's.

As a solution, this commit turns the TimerManager into a singleton.
The TimerManager member of DaemonCore is switched to a reference.