Avoid staging out unnecessary files. #4730
The blahp will create (and not cleanup) a file containing a limited
version of the job's proxy.  This file may be transferred back to
remote clients as part of the job sandbox (incorrectly - and
possibly in the clear!).  We should automatically exclude it.

Additionally, the .condor_pid_ns_status file is transferred back
as part of the job sandbox; automatically exclude that too.