Make graceful job removal independent of how job was removed. #4679
Whether a job was killed gracefully on hold or removal varied depending
on how it was held or removed. If the decision to hold/remove the job
came through the schedd, then a graceful killing was done. If a job
policy expression in the shadow caused the hold/removal, then a fast
kill was done.
Now, a graceful kill is done in both cases, controlled by config param
GRACEFULLY_REMOVE_JOBS and job attribute WantGracefulRemoval.