Update self over an existing UDP if shared_port is in use. #4850
If shared_port is in use and there's no specific command port, then
a UDP command socket is not created.  Hence, the collector ad is dropped
when trying to update itself over UDP.

As a self-update for a collector must go over the UDP (to prevent
blocking), we must switch updates from the normal command socket to
the superuser socket.