Allow direct and statistics queries.  #4843
This adds a new method to the python collector API - directQuery.
directQuery is equivalent to "condor_status -direct" in that it uses
the collector to find a given daemon, then requests an ad from the
daemon itself.

Additionally, we now allow the user to request statistics back from
query methods.

Finally, in testing this against a few remote collectors, I noticed
I missed a handful of ad types still in use; populate them in the enums.