Tolerate CMake being installed under a path containing spaces.
The Mac native package for CMake installs it in an Application
package named /Applications/CMake 2.8-12.app/ or similar.
The cmake command-line program (pointed to by CMAKE_COMMAND) is
under that directory.
We need to quote ${CMAKE_COMMAND} in our install() directives, else
it tries to run the program '/Applications/CMake', which fails.